Fix dynamic adaptive isolation of single crease patches.
The combination of dynamic isolation and single-crease patch optimizations
can introduce discontinuities in the limit surface along boundaries with
end-cap patches. The `SingleCreaseDynamicISolation` enum controls the
desired behavior of the limit surface as follows:

 - Smooth mode : local edge sharpness of single-crease patches is forcibly
   lowered in order to maintain continuity with the limit surface of
   neighboring end-cap patches. Effectively, the limit surface starts
   ignoring semi-sharp tags in areas where the level of feature isolation
   is dynamically lowered.

 - Sharp mode : edge sharpness of single-crease patches is always enforced,
   regardless of dynamic isolation. The consequence is that single-crease
   patches neighboring an end-cap patch from a lower level of isolation will
   cause a discontinuity in the limit surface. Semi-sharp edge tags are
   maintained, but neighboring dynamically isolated patches will show
   discontinuities.

note: this option is currently set at compile-time and can be set with
the definition of TMR_SINGLE_CREASE_DYNAMIC_ISOLATION. This decision may be
revisited if valid use-cases can be made for a run-time selection.

Fixes #13